Tasmota: Fix wrong param types for channels
The channels are "POWER", "POWER0", "POWER1" etc... so they're
strings, not integers.

We could also use integers, however, that would require special handling
for the single channel (only "POWER" without number) on every publish,
this way we can do that just once in the setup and be done with it.