somfytahoma: Use single param for identifying the gateway
The single unique identifier of Somfy Gateways is the Gateway PIN that
the device announces via zeroconf (txt + hostname). There's no point in
setting two params to the same value.

Note that due to a previous mistake this fixes also getting the IP
via zeroconf, because the old code was comparing against an incorrect
identifier.